pyproject.toml
[project]
name = "wattmeter-sharp"
version = "0.1.0"
description = "シャープの HEMS コントローラ JH-AG01 の内部通信を読み取って，電力センサーの計測値を取得するスクリプトです．"
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]
readme = "README.md"
requires-python = ">= 3.10"

dependencies = [
    "docopt-ng>=0.9.0",
    "fluent-logger>=0.11.1",
    "influxdb-client[ciso]>=1.44.0",
    "my-lib @ git+https://github.com/kimata/my-py-lib@4b9b8a47b15a3f0e8bd3e462b85e4f28ca523e1c",
    "zmq>=0.0.0",
    "pyserial>=3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sharp_hems"]

[tool.hatch.metadat]
allow-direct-references = true

[tool.rye]
managed = true
dev-dependencies = [
    "flaky>=3.8.1",
    "playwright>=1.45.1",
    "pytest-cov>=5.0.0",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.14.0",
    "pytest-playwright>=0.5.1",
    "pytest-xdist>=3.6.1",
    "time-machine>=2.15.0",
    "pytest>=8.3.2",
]

[tool.pytest.ini_options]
minversion = "6.0"

# addopts = "--verbose --log-cli-level=DEBUG --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html -vv"

addopts = "--verbose --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html -vv"

testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:The hookimpl CovPlugin.pytest_configure_node uses",
    "ignore:The hookimpl CovPlugin.pytest_testnodedown uses",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",

  # 以下，未使用関数
]

[tool.coverage.html]
directory = "tests/evidence/coverage"





# [tool.poetry.dependencies]
# PyYAML = "^6.0"
# coloredlogs = "^15.0.1"
# docopt = "^0.6.2"
# influxdb-client = {extras = ["ciso"], version = "^1.36.1"}
# fluent-logger = "^0.10.0"
# pyserial = "^3.5"
# slack-sdk = "^3.21.3"
# python = "^3.10"
# zmq = "^0.0.0"
